feat(seeding): complete all seeders for core models
- Added full seeding for departments, programs, lecturers, venues, and related entities
- Fixed bug in AuthenticationService allowing null department and collegeBuilding for DAPU users
- Corrected misplacement of code and name fields in Department and Venue seeders
- Cleaned up seeding logic for consistent data integrity

+++ src/main/java/com/henry/universitycourseschedular/models/_dto/OnboardUserDto.java
@@ -22,10 +22,10 @@ public record OnboardUserDto (
         String confirmPassword,
 
         @NotNull(message = "College building is required")
-        int collegeBuildingId,
+        Long collegeBuildingId,
 
         @NotNull(message = "Department is required")
-        int departmentId,
+        Long departmentId,
 
         @NotNull(message = "Invite Verification is required")
         Boolean inviteVerified
